requirements.txt
black @ git+https://github.com/psf/black@6ffc5f7b01eab341f1b2cd373b6faed0072a2351
click==8.1.3
Flask==2.2.3
flask-marshmallow==0.14.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
mypy-extensions==1.0.0
nose==1.3.7
packaging==23.0
pathspec==0.11.1
permissive-dict==1.0.4
platformdirs==3.1.1
postgres==4.0
psycopg2-binary==2.9.5
psycopg2-pool==1.1
PyJWT==2.6.0
six==1.16.0
SQLAlchemy==2.0.5.post1
tomli==2.0.1
typing-extensions==4.5.0
Werkzeug==2.2.3
zipp==3.15.0
